requirements.txt
steem
easysteem
python-binance
discord.py
